added initialization and reading of DAC and ADC on analog input/output threads. Message value is changed to float. ThreadRunner prints floats.

+++ RealTimeControlProject/ADC.c
@@ -124,7 +124,7 @@ void SetupAout()
 }
 
 // For generating a specific analog output at specified channel
-void GenerateAout( int voltage, int output_channel )
+void GenerateAout( double voltage, int output_channel )
 {
     unsigned int lsb_value = 0 ;
     unsigned int msb_and_channel = 0 ;
